fixed issue #232
Former-commit-id: 643d31ab7b09d1a9dda80beda6b7565ea37ae2d2

+++ cocos2d/Camera.m
@@ -34,7 +34,7 @@
 
 - (NSString*) description
 {
-	return [NSString stringWithFormat:@"<%@ = %08X | center = (%i,%i,%i)>", [self class], self, (int)centerX, (int)centerY,(int)centerZ];
+	return [NSString stringWithFormat:@"<%@ = %08X | center = (%.2f,%.2f,%.2f)>", [self class], self, centerX, centerY, centerZ];
 }
